Remove the kaggle api package
This breaks the twosigma competition because the code for both lives in a python package called `kaggle`.

The code for the twosigma competition under `kaggle.competitions` should probably be move.

Additionally, the kaggle api client package requires an API secret key. We don't offer a way to manage secrets for now and we don't want to encourage people adding their secret in the code (easy to leak if ever made public).