Remove vega3 package
Usage: 0 kernels per day.

Rational:

Unmaintained. Deprecated in favor of vega. Low GitHub stars, will consider add vega if > 1000 stars and if someone ask for it.

Some people are using Altair which requires vega to display the chart. It requires `vega` and not `vega3`. People are using the `pip install` workaround which is the preferred way to do for not very popular packages that are not required in KO competitions. See:
https://www.kaggle.com/general/63534