Remove tensorforce package.
Last usage in our platform was in October 2020.
In the past year, we have seen 14 notebooks using it only.

User wanting to use this package should use `!pip install tensorforce`
into their own kernel.

TensorForce 0.6.2 has dependencies on tensorflow==2.3.1 and
numpy==1.18.5 which is incompatible with our environment. We don't want
to constrain new release of TensorFlow for a package that is barely
being used. Hence, the reason for dropping it from our base image.

http://b/181584341