Uninstall google-cloud-bigquery-storage library.
This library is part of our base image (deeplearning-platform-release)
but not other libraries depend on it.

Removing this library ensures the bigquery library fall backs on using
regular BQ queries to fetch data.

The BigQuery storage APIs are not supported by our data proxy and we
don't monkey patch this client.

This change simply revert to the behavior we had with BQ 1.x libraries.

http://b/183041606