Fix TLS error when upgrading nvidia packages.
We used to download the nvidia packages using apt over http. However
this is not supported anymore causing a build failure.

I found that `apt` was failing to connect over https to the NVIDIA
download servers because of the `libgnutls.so` included in conda which
was used in place of the default version when we updated the
`LD_LIBRARY_PATH`. See the bug below for full context.

http://b/197989446