Install latest version of nbconvert to fix Jinja issue.
The latest version of Jinja changed the import for `Markup`.

The import was fixed in the latest version of `nbconvert`: https://github.com/jupyter/nbconvert/issues/1736

We include the version of `nbconvert` from the DLVM base image which is
released every ~2 weeks. However, the `nbconvert` fix was released only
6 hours ago.

Temporarily install thel latest version of `nbconvert`.

Add a test to prevent regression.

http://b/227194111