Move TORCH_VERSION ARG below the FROM (#1161)
This is necessary to make it available with that build context.

The other ARGs are above because they are used in the FROM statement themselves but not the build context of the last image.

I also updated the URL for installing `torch_xla`. The version of Torch should not include the minor version.