Update Colab Base image to colab_20250219-060225_RC01 (#1475)
We are upgrading the base image to the latest release image by colab:
colab_20250219-060225_RC01

Which includes the following upgrades:
TF 2.18
Python 3.11
Cuda 12.5

This PR includes a handful of fixes to resolve conflicts related to
these upgrade. Notably issues pertaining torch and cudnn.

We also bumped lightgbm version as well

We also included a fix to tune cli package conflict.